Alarm / MAJOR BUG - AlarmTimer and Timer instance. Does not work as intended.
Can successfully add an Alarm instance to the queue. When the Start button is pressed on the UI the Timer will start and when the Timer reaches zero the MessageBox notification will popup. Alerting the user that the time has elapsed.
[(A,10)] works and alarm is sounded correctly.

However adding multiple Alarms to the queue does not work as intended.
[(A,10), (B, 5)]
The alarms will both be added to the queue. However the Timer goes off at 5 seconds instead of the time at the Head item, which should be 10s.
This means that the Timer is only remembering the Last input or not the instance of the head.

If the timer is run again, after the Timer has reached zero. The alarm will sound at 0 seconds. This make me believe that the timer is set to zero at this point. And is not an instance of the Object - Unless we are setting that instnace time to zero as the timer gradually Countsdown.