AlarmModel / Fixed Alarm Timer Problem - Once countdown completion, and Alarm field equals zero, Alarm is removed.
The Alarms are now instantiated correctly.
Each Added item to the queue now has its own Alarm time.

The error where the Timer would only remember the last added Item is now fixed.
Each Alarm Timer can be counted down to zero, and if started again the Alarm is removed and the next Alarm will be started.

ToDo:
Update the UI as each Timer is removed.