Alarm Presenter / Added _presenter HeadTime function, and _view ViewCountdownNull
When an Alarm is added to the queue, the head Time is added to the UI. This makes the countdown appear to start at the correct value.
The function returns the Head.AlarmTime and send it to the View, ViewCountdownTime function.
Added exception handling to the HeadTime function.
If the head is null, it will call the _view ViewCountdownNull function to display that there is no time to Countdown in the view.