BugFix / Fixed UI countdown on multiple start button presses. On Countdow reaching Zero, the Head Item is removed from the Queue. Added Exception handling to Presenter ShowAlarms, StartAlarm, CountdownToUI_Tick
Fixed when pressing the start timer button multiple times, it would start multiple counters that would decrement the head time in the UI counter. This put the Head time and the UI counter out of sync and the UI counter would decrement extremely fast as it was counting down multiple times as well as not stopping resulting in a timer going into the negative numbers.  Eg, -100
Instead of decrementing each tick, we instead return the Head Time.

When the head time reaches zero, the Head of the queue is removed. This is done from the _presenter HeadCountdownTime and CountdownToUI_Tick event handler

Added exception handing for queue underflow when the last item in the queue was removed when the UI timer hit zero.