AlarmPresenter / Added RemoveAlarm & CountdownToUI_Tick logic
CountdownToUI:
Now checks if the _Counter is ticking, If so make the StartButton invalid. This stops spamming the start button which has a strange visual on the countdown. It does not harm the countdown but looks odd.
This also stop the _Counter going out of sync when a new Item is added Ahead of the Head of the Priority Queue.

RemoveAlarm:
If the _Counter is enabled and the RemoveAlarm function is called, stop the counter and make the StartButton interactable again.