AlarmModel / Modified AddAlarm Priority
Priority for newAlarms is now the timespan from tomorrow Midnight - DateTime.Now in seconds.

This means that each alarm is served in order of its creation time, the first alarm created is served first and the second second, etc. All based on the time difference.

The longer the time difference between midnight and now will have the highest priority based on the SortedArrayPriorityQueue where the largest Priority integer has the highest priority.

Changed from AlarmTime to Timespan timeDifference