ClockView / Added alarmView creation logic in ButtonOpenForm2 event and added OnAlarmViewFormClosing eventhandler *Important
ButtonOpenForm2 :
If there is all ready an alarmView that exists OR if alarmView has been disposed, create a new AlarmView, a new Model and a new Presenter.
Subscribe to the OnAlarmViewFormClosing event then show the new alarmView.

OnAlarmViewFormClosing:
Eventhandler to handle the FormClosing event. This stops the form from being Closed (where we would lose the state of our alarmView, Model, Presenter and the Alarms that have all ready been created or being counteddown).
Instead of the form being Closed it is now being Hidden when the X button on the alarmView is pressed.
This allows us to keep the state of our alarmView and our created alarms.

IView:
Modified the FormClosingEventHander to properly allow for Hidden forms and the event to be triggered. Previously it was FormClosedEvent which would NEVER be triggerd due to the form being in a Hidden state (and not Closed).