Wiring Up AlarmPresenter to ClockPresenter / Modified ClockView to create the AlarmPresenter with a reference of ClockPresenter in its constructor, to allow for passing information back
Created:
AlarmModel - ShowHeadName
AlarmModel - ShowHeadTimeIntSeconds

Test Methods:
ClockPresenter - CanYouSeeThisName
ClockPresenter - CanYouSeeThisIntTime
(Need to combine these metods then convert the string Time to TimeSpan in the ClockPresenter, to pass this to the ClockView Alarm time UI.

Modified:
ClockView - ButtonOpenForm2_Click method that creates the AlarmMVP.  Allows for the Presenter to be created with a reference of the ClockPresenter to pass information back.

ToDo:
Combine the test methods to send AlarmName and AlarmTime to the ClockPresenter, in one function. Then convert the time to TimeSpan from its string value