AlarmPresenter / Added UpdateClockViewWithNextAlarm method in AlarmPresenter and added error handling to display when there was no alarms in the ClockView label. Modified the Add and Remove Methods in AlarmPresenter
Created a new method to get the Head name and time (timespan) and send that to the ClockPresenter to update the ClockView UI label with the next alarm info.
If there was no next alarm, call the method in the ClockPresenter to output that to the ClockView UI label.

Removed redundant test methods that were combined into one.

Modified the AlarmPresenter Add and Remove methods to call the new UpdateClockViewWithNextAlarm method.